- PotentiallyIrrationalNum must provide one of monzo, ratio, or decimal, and CentsPosition therefore does not have to provide cents, since it can rely on one of the others being present, the goal being for pitches to be able to rely more on Num-based methods; however Pitch in general now can't rely on one of monzo, ratio, or decimal being present since CentsPosition can have only cents. so I really think we won't have truly accomplished something great until we do decide to force Pitch to not even include cents at all (so I agree that cents is a valuable way of representing pitches, but the thing is, that unlike decimal, monzo, and ratio, they're just not in the same sense a representation of the SAME NUMBER. even the up-and-coming ED representation will be the same number truly. so cents should just be a thing you convert into and out of). that way you can simplify a lot hopefully and maybe even entirely eliminate the CentsPosition and JiPitch types. okay so I temporarily made all the BoundPositions enforce that they do have cents for sure, so I forestalled needing to switch them from cents to decimal. but before I did that, I halfway started down the path of forcing decimals to be there, so I put them on the ina midpoints, comma means, and size category bounds, and the tests confirm that they're there, but I haven't yet forced the cents representations to go away.

+++ src/scripts/jiPitch/analyzeJiPitch/parse.ts
@@ -10,10 +10,6 @@ const parseJiPitch = (): JiPitch => {
     if (jiPitchText) {
         const pitch = parsePitch(jiPitchText)
 
-        // todo: DECIMAL & CENTS
-        //  it is possible here that parsePitch can return a pitch which has only cents.
-        //  without one of monzo, ratio, or decimal, which won't work for jiPitch.
-        //  but that wouldn't be the case if we fix the problem.
         if (!isUndefined(pitch.monzo) || !isUndefined(pitch.ratio) || !isUndefined(pitch.decimal)) {
             jiPitch = pitch as JiPitch
         } else {
@@ -27,8 +23,8 @@ const parseJiPitch = (): JiPitch => {
         jiPitch = { ratio: program.ratio }
     } else if (program.commaName) {
         jiPitch = { monzo: program.commaName }
-    } else if (program.decimal) {
-        jiPitch = { decimal: program.decimal }
+    } else if (program.integer) {
+        jiPitch = { decimal: program.integer }
     } else {
         throw new Error("Unable to parse JI pitch.")
     }
